-implemented appoinmentDataentryservice for Ui interaction for appointments
